got deletion from grouplist working
