+++ .gitignore
@@ -6,3 +6,4 @@ db.sqlite3
 /assets/bundles/
 node_modules
 webpack-stats.json
+
