+++ webpack-stats.json
@@ -1,1 +1,1 @@
-{"status":"done","chunks":{"main":[{"name":"main-b162c61d7783ac7d5ff2.js","path":"/home/pb4301/Documents/git/mailer_application/assets/bundles/main-b162c61d7783ac7d5ff2.js"}]}}
+{"status":"done","chunks":{"main":[{"name":"main-83d5fda783f9cbe9a7e1.js","path":"/home/pb4301/Documents/git/mailer_application/assets/bundles/main-83d5fda783f9cbe9a7e1.js"}]}}
